requirements.txt
asgiref==3.8.1
attrs==23.2.0
autopep8==1.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.1
distlib==0.3.8
dj-stripe==2.8.4
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.2
django-formtools==2.3
django-ratings==0.3.7
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
filelock==3.13.3
idna==3.7
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.8.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
simplejson==3.19.2
sqlparse==0.4.2
stripe==4.2.0
toml==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
